# apps/server/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: SiteCRM API
  version: 0.2.0
  description: |
    Primary REST API for SiteCRM.  Handles authentication, lead management,
    outreach, notifications, site generation, and deployments.

    **Authentication:** Most endpoints require a Bearer JWT.  Obtain a token
    via `POST /api/auth/login` or `POST /api/auth/register` and pass it as
    `Authorization: Bearer <token>` on subsequent requests.

    **Common error responses (apply to every operation unless noted):**
    - `400 Bad Request` — request body or query failed validation
      (`additionalProperties: false` rejects unknown fields).
    - `401 Unauthorized` — missing, invalid, or expired token on a protected route.
    - `503 Service Unavailable` — the database is not configured/reachable.

    All errors use the shared `Error` schema. Reusable definitions are under
    `components/responses` (`BadRequest`, `Unauthorized`, `NotFound`,
    `ServiceUnavailable`).
  contact:
    name: Riverfront Labs
servers:
  - url: http://api.localhost
    description: Local development (Traefik)
  - url: http://localhost:3000
    description: Direct local development
...
paths:
...
  /api/auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Create a new account
      description: |
        Creates a user record with a bcrypt-hashed password (cost 12) and
        returns a signed JWT plus the public user object.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Sign in to an existing account
      description: |
        Verifies credentials and returns a signed JWT. Returns `401` for any
        mismatch — intentionally does not reveal whether the email exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current signed-in user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Leads ────────────────────────────────────────────────────────────────────
  /api/leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads with optional filters
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
        - name: status
          in: query
          description: Filter by one or more statuses (repeat the param for multiple).
          schema:
            oneOf:
              - $ref: '#/components/schemas/LeadStatus'
              - type: array
                items:
                  $ref: '#/components/schemas/LeadStatus'
        - name: shortlisted
          in: query
          schema: { type: boolean }
        - name: search
          in: query
          description: Case-insensitive substring match on name, email, or location.
          schema: { type: string, maxLength: 200 }
      responses:
        '200':
          description: Paginated lead list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLeads'

    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLead'
      responses:
        '201':
          description: Lead created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Reusable error responses — reference from operations via
  # `$ref: '#/components/responses/<Name>'`. See the info description for when
  # each applies across the API.
...
  schemas:
...
    User:
      type: object
      required: [id, email, firstName, lastName, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: Stored as bcrypt hash (cost 12). Max 72 chars — bcrypt truncates beyond that.
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: 'Signed JWT — include as `Authorization: Bearer <token>`'
        user:
          $ref: '#/components/schemas/User'

    # ── Leads ─────────────────────────────────────────────────────────────────
    LeadStatus:
      type: string
      enum: [new, contacted, qualified, proposal, won, lost]
...
    Lead:
      type: object
      required: [id, userId, name, status, shortlisted, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: ['string', 'null']
          format: email
        website:
          type: ['string', 'null']
        facebookPageId:
          type: ['string', 'null']
          description: Facebook Page ID for Messenger outreach.
        instagramAccountId:
          type: ['string', 'null']
          description: Instagram Business Account ID for DM outreach.
        websiteScore:
          type: ['number', 'null']
          description: 0–100 score from the intelligence website analyser.
        websiteGrade:
          type: ['string', 'null']
        websiteNotes:
          type: ['string', 'null']
        score:
          type: ['number', 'null']
          description: Composite intelligence score (0–100).
        rating:
          type: ['number', 'null']
          description: Google Places star rating (1–5).
        reviews:
          type: ['integer', 'null']
        type:
          type: ['string', 'null']
          description: Business category (e.g. "restaurant", "law firm").
        location:
          type: ['string', 'null']
        placeId:
          type: ['string', 'null']
          description: Google Places ID.
        linkedSiteId:
          type: ['string', 'null']
          format: uuid
        status:
          $ref: '#/components/schemas/LeadStatus'
        shortlisted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NewLead:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: string
          format: email
          maxLength: 255
        website:
          type: string
          maxLength: 512
        facebookPageId:
          type: string
          maxLength: 255
        instagramAccountId:
          type: string
          maxLength: 255
        type:
          type: string
          maxLength: 100
        location:
          type: string
          maxLength: 255
        placeId:
          type: string
          maxLength: 255
        status:
          $ref: '#/components/schemas/LeadStatus'
        shortlisted:
          type: boolean
...
    PaginatedLeads:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Lead'
        total:
          type: integer
          description: Total matching leads across all pages.
        page:
          type: integer
        pageSize:
          type: integer
...
    Error:
      type: object
      required: [statusCode, error, message]
      description: Standard error envelope — matches Fastify's default error shape.
      properties:
        statusCode:
          type: integer
        error:
          type: string
          description: HTTP reason phrase, e.g. `Not Found`
        message:
          type: string
          description: Human-readable explanation of what went wrong